Исправление ошибок. Доработка
