+++ src/utils/common.js
@@ -1,0 +1,21 @@
+export const getRandomInteger = (a = 0, b = 1) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+
+  return Math.floor(lower + Math.random() * (upper - lower + 1));
+};
+
+export const getRandomArrayElement = (arr) =>
+  arr[getRandomInteger(0, arr.length - 1)];
+
+export const updateItem = (items, update) => {
+  const index = items.findIndex((item) => item.id === update.id);
+
+  if (index === -1) {
+    return items;
+  }
+
+  return [...items.slice(0, index), update, ...items.slice(index + 1)];
+};
+
+export const isEscapeKey = (evt) => evt.key === 'Escape' || evt.key === 'Esc';
