+++ src/view/sort-view.js
@@ -1,16 +1,16 @@
 import AbstractView from '../framework/view/abstract-view.js';
 
 export const SortType = {
-  DAY: 'day',
+  DEFAULT: 'day',
   EVENT: 'event',
   TIME: 'time',
   PRICE: 'price',
-  OFFER: 'offer'
+  OFFER: 'offer',
 };
 
-const createSortTemplate = () => (
+const createSortTemplate = () =>
   `<form class="trip-events__trip-sort  trip-sort" action="#" method="get">
-   ${Object.values(SortType).map((item) => (
+   ${Object.values(SortType).map((item) =>
     `<div class="trip-sort__item  trip-sort__item--${item}">
         <input
           id="sort-${item}" 
@@ -23,11 +23,11 @@ const createSortTemplate = () => (
           ${item === 'offer' ? 'disabled' : ''}
         >
         <label class="trip-sort__btn" for="sort-${item}">${item}</label>
-      </></div>`)).join('')}
-  </form>`
-);
+      </></div>`)
+    .join('')}
+  </form>`;
 
-export default class SortView extends AbstractView{
+export default class SortView extends AbstractView {
   get template() {
     return createSortTemplate();
   }
